Introducing InstanceProcessor
